Allow User Interaction on tag's label
UI tests are unable to tap the tag views via the accessibility identifier because the TTGTextTagLabel does not have an identifier - the gradient label within it does. However, the gradient label is currently not tappable. This change makes the label tappable so that UI tests can find the TTGTextTagLabel via the gradient label and tap it.